Add a more strict test to check if the rank is a float and favour 0 as a fallback instead of 0.0.

+++ spec/textacular/searchable_spec.rb
@@ -68,7 +68,7 @@ class SearchableTest < Test::Unit::TestCase
         qcont_with_author = @qcont.becomes(WebComicWithSearchableNameAndAuthor)
         search_result = WebComicWithSearchableNameAndAuthor.fuzzy_search('Questio')
         assert_equal [qcont_with_author], search_result
-        assert_kind_of Numeric, search_result.first.attributes.find { |k, _| k[0..3] == 'rank' }.last
+        assert_kind_of Float, search_result.first.attributes.find { |k, _| k[0..3] == 'rank' }.last
       end
 
       should "define :searchable_columns as private" do
